+++ backend/model/User.js
@@ -1,0 +1,22 @@
+const mongoose = require("mongoose");
+
+const UserSchema = mongoose.Schema({
+  username: {
+    type: String,
+    required: true
+  },
+  email: {
+    type: String,
+    required: true
+  },
+  password: {
+    type: String,
+    required: true
+  },
+  createdAt: {
+    type: Date,
+    default: Date.now()
+  }
+});
+
+module.exports = mongoose.model("users", UserSchema);
