--- src/components/header.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Header = () => {
-    return (
-        <header className="header">
-            <h1>Weather-Finder App</h1>
-        </header>
-    )
-}
-
-export default Header;
